feat: move into confition packets and fix ITs
